Allow use of last_error! without extra library calls
Systematic get_last_error() call even when the wrapped call did not
return an error looks wasteful.  This change allows to tell the macro
that the call has succeeded, and skip the unneeded call.

Did not apply to place that require more insights:
- domain_hvm_getcontext: errors are checked for 2nd call only, and then
  only after additional unsafe code


I have the same feeling towards systematic clear_last_error(): my gut
feeling is that this *should* only be useful when calling
get_last_error() after the wrapped call succeeds, as on error we're
pretty sure `last_error` will be overwritten.  That possibly remains
to be double-checked, but situation where it does not hold are likely
to be bugs?